refactor(events): consolidate EventPublicRsvpSection into EventMemberSection
EventPublicRsvpSection duplicated a read-only subset of EventMemberSection
now that both land on the same event detail page (Issue 854). Merge them
into one component: EventMemberSection now accepts an optional token prop
and renders for both real members and token-holding non-members, with
host/invite/admin affordances gated on a real auth session (never the
token) rather than on which component is rendered.

- Extract HostSection (+ chips) into EventHostSection.tsx, and the shared
  Card primitive into EventDetailCard.tsx, to keep files under the
  project's line limit while adding the token branch.
- EventDetailScreen renders EventMemberSection for both the authed and
  token-unlocked branches instead of two parallel components.
- Delete EventPublicRsvpSection.tsx and its test; add token-holder
  coverage to EventMemberSection.test.tsx asserting hosts, invite,
  email blast, group text, and admin actions never render for a token
  holder, while location/rsvp/comments still do.
